third_party/blink/web_tests/platform/generic/fast/css/getComputedStyle/computed-style-properties-expected.txt

outline
content
counter
Heading A
Subheading A
Subheading B
Heading B
PASS computedStyleFor('outline', null, 'outline-offset') is '5px'
PASS computedStyleFor('content', 'before', 'content') is "\"text\""
PASS computedStyleFor('content', 'after', 'content') is "\"test \" url(\"data:image/gif;base64,R0lGODlhAQABAJAAAP8AAAAAACwAAAAAAQABAAACAgQBADs=\")"
PASS computedStyleFor('counter', null, 'counter-reset') is 'section 0'
PASS str.indexOf('subsection 0') != -1 is true
PASS str.indexOf('anothercounter 5') != -1 is true
PASS computedStyleFor('counter1', 'before', 'counter-increment') is 'section 1'
PASS computedStyleFor('subcounter2', 'before', 'counter-increment') is 'subsection 1'
PASS computedStyleFor('subcounter2', 'before', 'content') is "counter(section) \".\" counter(subsection) \". \""
PASS successfullyParsed is true

TEST COMPLETE
